shop_backend/requirements.txt
asgiref==3.8.1
contourpy==1.3.3
cycler==0.12.1
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
django-jalali-date==2.0.0
django-render-partial==0.4
filelock==3.17.0
fonttools==4.59.1
jalali_core==1.0.0
jdatetime==5.2.0
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pillow==11.2.1
platformdirs==4.3.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.29.2
virtualenvwrapper-win==1.2.7
djangorestframework==3.15.2